Remove airplane.ico resource from project
The `BroadcastPluginSDK.csproj` file was updated to remove the inclusion of the `airplane.ico` resource. The corresponding `<None Update="Resources\airplane.ico">` entry and its `<CopyToOutputDirectory>` directive were deleted. Additionally, the `airplane.ico` binary file was removed from the project.

The `airplane.png` resource remains in the project and continues to be included in the package and copied to the output directory, potentially replacing the functionality of the removed `airplane.ico` file.